ti816x: Adjust device voltage based on characterization data
The patch adjusts the voltage fed to TI816X chip from TPS40041 based
on the characterization data for each device. This will reduce the
power consumption and thermal dissisipation of the device.

To enable this feature, enable CONFIG_TI816X_VOLT_SCALE in configuration.
A parameter table containing the mapping of characterized data in efuse to
GPIO values need to be obtained from the hardware team and populated in
software.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>