OMAP3: Add MMC SPL support to omap3_evm
Add Hynix 200MHz timing information to <asm/arch-omap3/mem.h>.  We
don't calculate the MCFG value as we have a special value coming from
x-loader to use here, so it's omitted from the header file.  In order
to determine if we have Hynix or Micron memory we have to see what
NAND chip we are running on.  Tested on a rev G board.

TODO: Support NAND booting.

Signed-off-by: Tom Rini <trini@ti.com>