am335x_evm: single-byte address EEPROM for board identifier
The BeagleBone has an EEPROM that uses a single byte address instead of a
2-byte address used by the EVM versions.  This simply adds a quick attempt
at reading the board identifier header with a 1-byte address if the header
signature isn't correct using a 2-byte address.

This is only build tested.  Having a script to program the expected EEPROM
contents would help me reduce my time to test this.